lw_reviews/src/website/requirements/project.txt
--extra-index-url=http://dist.gondor.io/dev/
gondor==1.0b1.post10
Django==1.3
PyYAML==3.09
distribute==0.6.15
django-extensions==0.6
django-piston==0.2.2
django-taggit==0.9.3
django-tastypie==0.9.9
django-templatetag-sugar==0.1
eventlet==0.9.14
greenlet==0.3.1
gunicorn==0.12.1
mimeparse==0.1.3
nashvegas==0.6.3
psycopg2==2.4
python-dateutil==1.5
python-digest==1.7